requirements.txt
Flask
gunicorn
Flask-SQLAlchemy
Flask-Migrate
Flask-Bcrypt
Flask-JWT-Extended
python-dotenv
